@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -27,19 +27,23 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  awesome_notifications: any
   bottom_drawer: ^0.0.5
+  connectivity_plus: ^2.3.6
   crypto: ^3.0.2
   cupertino_icons: ^1.0.2
   device_info_plus: ^4.0.0
   diff_match_patch: ^0.4.1
   dio: ^4.0.6
+  disable_battery_optimization: ^1.1.0+1
+  enum_flag: ^1.0.1
   equatable: ^2.0.3
   flutter:
     sdk: flutter
   flutter_background_service: ^2.1.3
   flutter_bloc: ^8.0.1
+  flutter_local_notifications: ^9.7.0
   flutter_svg: ^1.0.3
+  flutter_translate: ^4.0.3
   fuzzy: ^0.4.0-nullsafety.0
   hive: ^2.1.0
   hive_flutter: ^1.1.0
@@ -49,6 +53,7 @@ dependencies:
   package_info_plus: ^1.4.2
   quick_notify: ^0.2.1
   settings_ui: ^2.0.2
+  shared_preferences: ^2.0.15
   system_tray: ^2.0.1
   throttling: ^1.0.0
   universal_html: ^2.0.8
@@ -57,6 +62,7 @@ dependencies:
   url_strategy: ^0.2.0
   validators: ^3.0.0
   window_manager: ^0.2.5
+  workmanager: ^0.5.0
 
 dev_dependencies:
   bloc_test: ^9.0.3
@@ -83,6 +89,7 @@ flutter:
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
   assets:
+    - assets/i18n/
     - assets/resources/logos/svg/
     - assets/resources/favicon.ico
     - assets/resources/logos/png/Round_colored_logo_110px.png
