@@ -1,3 +1,4 @@
+import 'package:disable_battery_optimization/disable_battery_optimization.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -5,46 +6,91 @@ import 'package:flutter_memosync/src/authentication/authentication.dart';
 import 'package:flutter_memosync/src/home/home.dart';
 import 'package:flutter_memosync/src/home/repositories/memo.dart';
 import 'package:flutter_memosync/src/login/views/login_page.dart';
-import 'package:flutter_memosync/src/services/desktop_window_manager.dart';
+import 'package:flutter_memosync/src/services/background_handlers/desktop_window_manager.dart';
 import 'package:flutter_memosync/src/services/logger.dart';
 import 'package:flutter_memosync/src/services/models/settings.dart';
+import 'package:flutter_memosync/src/services/notification_service.dart';
 import 'package:flutter_memosync/src/services/repositories/user.dart';
 import 'package:flutter_memosync/src/services/storage/storage.dart';
 import 'package:flutter_memosync/src/splash/splash.dart';
 import 'package:flutter_memosync/src/widgets/route_404.dart';
-import 'package:hive_flutter/hive_flutter.dart';
+import 'package:flutter_translate/flutter_translate.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:universal_platform/universal_platform.dart';
 
 /// Widget providing the [MaterialApp], listening for changes on the
 /// [AuthenticationBloc] for routing.
 class AppView extends StatefulWidget {
   /// Default constructor
-  const AppView({Key? key}) : super(key: key);
+  const AppView({super.key});
 
   @override
   State<AppView> createState() => _AppViewState();
 }
 
-class _AppViewState extends State<AppView> {
+class _AppViewState extends State<AppView> with WidgetsBindingObserver {
   final _navigatorKey = GlobalKey<NavigatorState>();
 
   NavigatorState get _navigator => _navigatorKey.currentState!;
 
-  bool dark = true;
   bool startup = true;
+  SharedPreferences? _sharedPreferences;
+
+  @override
+  void initState() {
+    super.initState();
+    WidgetsBinding.instance.addObserver(this);
+    Future.microtask(() async {
+      _sharedPreferences = await SharedPreferences.getInstance();
+    });
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    super.dispose();
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    App.lifeCycleState = state;
+    _sharedPreferences?.setString('lifeCycleState', state.name);
+  }
 
   @override
   Widget build(BuildContext context) {
-    return ValueListenableBuilder<Box<SettingsObject>>(
-      valueListenable: Storage.settingsStorageStream,
-      builder: (context, settingsBox, _) {
-        final settings = settingsBox.get('settings');
+    final localizationDelegate = LocalizedApp.of(context).delegate;
+
+    return ValueListenableBuilder<SettingsObject>(
+      valueListenable: Storage.settingsStorageStream(),
+      builder: (context, settings, _) {
         return MaterialApp(
           theme: App.lightTheme,
           darkTheme: App.darkTheme,
-          themeMode:
-              (settings?.darkMode ?? true) ? ThemeMode.dark : ThemeMode.light,
+          themeMode: (settings.darkMode) ? ThemeMode.dark : ThemeMode.light,
           navigatorKey: _navigatorKey,
+          localizationsDelegates: [
+            localizationDelegate,
+          ],
+          supportedLocales: localizationDelegate.supportedLocales,
+          locale: localizationDelegate.currentLocale,
           builder: (context, child) {
+            Future(() async {
+              if (!UniversalPlatform.isDesktopOrWeb &&
+                  !(await DisableBatteryOptimization
+                          .isAllBatteryOptimizationDisabled ??
+                      false)) {
+                await DisableBatteryOptimization
+                    .showDisableAllOptimizationsSettings(
+                  'Enable Auto Start',
+                  'Follow the steps and enable the auto start of this app',
+                  'Your device has additional battery optimization',
+                  '''
+Follow the steps and disable the optimizations to allow smooth functioning of this app''',
+                );
+              }
+            });
+
             Logger.info('built');
             return BlocListener<AuthenticationBloc, AuthenticationState>(
               listenWhen: (previous, current) =>
@@ -58,6 +104,7 @@ class _AppViewState extends State<AppView> {
                   );
                   switch (state.status) {
                     case AuthenticationStatus.authenticated:
+                      NotificationService.initNotifications();
                       _navigator.pushAndRemoveUntil<void>(
                         HomePage.route(),
                         (route) => false,
@@ -126,11 +173,11 @@ class _AppViewState extends State<AppView> {
 class App extends StatelessWidget {
   /// Default constructor requiring both repositories
   const App({
-    Key? key,
+    super.key,
     required this.authenticationRepository,
     required this.userRepository,
     required this.memoRepository,
-  }) : super(key: key);
+  });
 
   /// Repository used to handle authentication calls
   final AuthenticationRepository authenticationRepository;
@@ -146,6 +193,9 @@ class App extends StatelessWidget {
 
   /// Used for Logger.graphic
   static BuildContext? context;
+
+  /// Used to access life cycle state from anywhere in the app
+  static AppLifecycleState? lifeCycleState;
 
   /// Light theme used in the app
   static final lightTheme = ThemeData(
@@ -170,21 +220,24 @@ class App extends StatelessWidget {
     App.context = context;
     Logger.info('Launching App');
     return DesktopWindowManager.windowWrapper(
-      child: MultiRepositoryProvider(
-        providers: [
-          RepositoryProvider.value(
-            value: authenticationRepository,
+      child: LocalizationProvider(
+        state: LocalizationProvider.of(context).state,
+        child: MultiRepositoryProvider(
+          providers: [
+            RepositoryProvider.value(
+              value: authenticationRepository,
+            ),
+            RepositoryProvider.value(
+              value: memoRepository,
+            ),
+          ],
+          child: BlocProvider(
+            create: (_) => AuthenticationBloc(
+              authenticationRepository: authenticationRepository,
+              userRepository: userRepository,
+            ),
+            child: const AppView(),
           ),
-          RepositoryProvider.value(
-            value: memoRepository,
-          ),
-        ],
-        child: BlocProvider(
-          create: (_) => AuthenticationBloc(
-            authenticationRepository: authenticationRepository,
-            userRepository: userRepository,
-          ),
-          child: const AppView(),
         ),
       ),
     );
