@@ -9,13 +9,12 @@
 import 'dart:typed_data';
 
 import 'package:flat_buffers/flat_buffers.dart' as fb;
+import 'package:flutter_memosync/src/services/models/memo.dart';
+import 'package:flutter_memosync/src/services/models/settings.dart';
+import 'package:flutter_memosync/src/services/models/user.dart';
 import 'package:objectbox/internal.dart'; // generated code can access "internal" functionality
 import 'package:objectbox/objectbox.dart';
 import 'package:objectbox_flutter_libs/objectbox_flutter_libs.dart';
-
-import 'src/services/models/memo.dart';
-import 'src/services/models/settings.dart';
-import 'src/services/models/user.dart';
 
 export 'package:objectbox/objectbox.dart'; // so that callers only have to import this file
 
